requirements.txt
django==5.0.6
python-dateutil==2.9.0